+++ src/components/mypage/Mypage.tsx
@@ -8,9 +8,10 @@ import MypagePost from './MypagePost';
 
 export default function Mypage() {
   const [activeTab, setActiveTab] = useState('작성글');
+  const [searchKeyword, setSearchKeyword] = useState('');
 
   return (
-    <section className="w-full bg-[var(--gray-40)] px-48 pt-6 pb-15">
+    <section className="w-full h-[650px] bg-[var(--gray-40)] px-48 pt-6">
       <div className="flex justify-between items-center mb-10">
         {/* 메뉴 */}
         <ul className="flex gap-8 text-lg text-[var(--gray-300)] font-medium">
@@ -34,6 +35,8 @@ export default function Mypage() {
           <Input
             className="w-[250px] bg-[var(--white)] border-[var(--white)] pr-10"
             placeholder="검색어를 입력해주세요"
+            value={searchKeyword}
+            onChange={(e) => setSearchKeyword(e.target.value)}
           />
           <Search
             size={16}
@@ -43,13 +46,13 @@ export default function Mypage() {
         </div>
       </div>
 
-      {/* 카드 or 테이블 */}
-      <div className="mt-4">
+      <div className="mt-4 h-[calc(100%-80px)] overflow-y-auto pr-2">
         {activeTab === '레이어' ? (
-          <MypageLayer />
+          <MypageLayer searchKeyword={searchKeyword} />
         ) : (
           <MypagePost
             activeTab={activeTab as '작성글' | '참여글' | '좋아요글'}
+            searchKeyword={searchKeyword}
           />
         )}
       </div>
