+++ src/components/mypage/profiletab/AchievementTab.tsx
@@ -1,0 +1,7 @@
+export default function AchievementTab() {
+  return (
+    <div className="text-[15px] text-[var(--gray-400)] mt-4">
+      업적 보기 영역 (추후 구현)
+    </div>
+  );
+}
