+++ src/components/mypage/profiletab/InterestTab.tsx
@@ -1,0 +1,7 @@
+export default function InterestTab() {
+  return (
+    <div className="text-[15px] text-[var(--gray-400)] mt-4">
+      관심 분야 설정 영역 (추후 구현)
+    </div>
+  );
+}
